+++ examples/dreamcast/raylib/tetris/src/constants/vmuIcons.h
@@ -1,0 +1,17 @@
+/* KallistiOS ##version##
+
+   examples/dreamcast/raylib/tetris/src/constants/vmuIcons.h
+   Copyright (C) 2024 Cole Hall
+
+*/
+
+#pragma once
+
+extern const char *vmuI;
+extern const char *vmuO;
+extern const char *vmuT;
+extern const char *vmuL;
+extern const char *vmuJ;
+extern const char *vmuS;
+extern const char *vmuZ;
+extern const char *vmuNULL;